requirements.txt
asgiref==3.5.0
Django==4.0.3
django-extra-views==0.14.0
Pillow==9.1.0
sqlparse==0.4.2
django-debug-toolbar